jobs/ml-ops/job/requirements.txt
pandas==2.0.2
numpy==1.24.3
scikit-learn==1.2.2
imblearn==0.0
matplotlib==3.7.1
boto3==1.29.3
python-dotenv==1.0.0